vstmac版とりあえず動作
git-svn-id: http://picopicose.com/svn/C700@81 38616c85-8a89-495d-9829-d15a903298fa

+++ C700Edit.h
@@ -37,7 +37,7 @@ public:
 	virtual bool	open(void *ptr);
 	virtual void	close();
 	
-	virtual void	setParameter(long index, float value);
+	virtual void	setParameter(int index, float value);
 	virtual void	valueChanged(CControl *pControl);	
 	long			getTag();
